03_nyc_citi_bike_streaming/requirements.txt
psycopg2-binary==2.9.1
requests
kafka-python